async_hooks: improve resource stack performance
Removes some of the performance overhead that came with
`executionAsyncResource()` by storing async resources that
are managed by JS and those managed by C++ separately, and
instead caching the result of `executionAsyncResource()` with
low overhead to avoid multiple calls into C++. In particular,
this is useful when using the async_hooks callback trampoline.

This particularly improves performance when async hooks are not
being used.

(This is continuation of https://github.com/nodejs/node/pull/33575.)

Refs: https://github.com/nodejs/node/pull/33575